+++ src_new/aliases/levi_ataxia/for_levi/levi_062424/201_Sysupdate.lua
@@ -1,0 +1,15 @@
+--[[mudlet
+type: alias
+name: Sysupdate
+hierarchy:
+- Levi_Ataxia
+- General
+attributes:
+  isActive: 'yes'
+  isFolder: 'no'
+regex: ^sysupdate$
+command: ''
+packageName: ''
+]]--
+
+ataxia.updater.doUpdate()